promote floats to double in tillotson eos

+++ sph+nln/eos.c
@@ -108,26 +108,26 @@ double murnaghan_eos(double k_bulk, double n_M, double eta) {
     m->cweib = 0.0;
 }
 
-void tillotson_eos (float rho, float u, Material_t *m, float *pressure, float *cs) {
-	float PC = 0.; /* pressure, current */
-	float csC = 0.; /* sound speed, current */
-	float rho0m1 = 1. / m->rho0;
-	float rhom1 = 1. / rho;
-	float eta = rho * rho0m1;
-	float mu = eta - 1.;
-	float csmin = 0.25 * m->A * rho0m1;
-	float Pmin = 0.; /* set floor on pressure */
-	float c1 = u / (m->u0 * eta * eta);
-	float c2 = 1. / (c1 + 1.);
+void tillotson_eos (double rho, double u, Material_t *m, double *pressure, double *cs) {
+	double PC = 0.; /* pressure, current */
+	double csC = 0.; /* sound speed, current */
+	double rho0m1 = 1. / m->rho0;
+	double rhom1 = 1. / rho;
+	double eta = rho * rho0m1;
+	double mu = eta - 1.;
+	double csmin = 0.25 * m->A * rho0m1;
+	double Pmin = 0.; /* set floor on pressure */
+	double c1 = u / (m->u0 * eta * eta);
+	double c2 = 1. / (c1 + 1.);
 
 	/* A. Brundage, 2013, 12th Hypervelocity Impact Symposium, 
 	 * Procedia Engineering, 58, 461-470 describes an implementation of this eos - CIE */
 	if (u > m->Eiv && eta < 1.) {
 		/* eq. 3, but there u> u_cv */
-		float d1 = m->rho0 * rhom1; /* = 1/eta ? */
-		float d2 = d1 - 1.;
-		float ex1 = (m->beta * d2 < 60) ? exp (- m->beta * d2) : 0.0;
-		float ex2 = (m->alpha * d2 * d2 < 60) ? exp (-m->alpha * d2 * d2) : 0.0;
+		double d1 = m->rho0 * rhom1; /* = 1/eta ? */
+		double d2 = d1 - 1.;
+		double ex1 = (m->beta * d2 < 60) ? exp (- m->beta * d2) : 0.0;
+		double ex2 = (m->alpha * d2 * d2 < 60) ? exp (-m->alpha * d2 * d2) : 0.0;
 		*pressure = m->a * rho * u;
 		*pressure += ex2 * (m->b * rho * u * c2 + m->A * mu * ex1);
 
@@ -147,9 +147,9 @@ void tillotson_eos (float rho, float u, Material_t *m, float *pressure, float *c
 		csC += PC * rhom1 * (m->a + m->b * c2 * c2);
 
 		if (u > m->Eiv && u < m->Ecv && eta < 1) {
-			float e1 = m->Ecv - u;
-			float e2 = u - m->Eiv;
-			float e3 = 1. / (m->Ecv - m->Eiv);
+			double e1 = m->Ecv - u;
+			double e2 = u - m->Eiv;
+			double e3 = 1. / (m->Ecv - m->Eiv);
 			/* eq. 5 */
 			*pressure = (e2 * *pressure + e1 * PC) * e3;
 			*cs = (e2 * *cs + e1 * csC) * e3;
